+++ cmeel/utils.py
@@ -111,7 +111,7 @@ def get_tag(pyproject) -> str:
 
     if deprecate_build_system(pyproject, "py3-none", False):
         warnings.warn(
-            "The 'py3-none = true' key is deprecated. Please use 'has_sitelib = false'",
+            "The 'py3-none = true' key is deprecated. Please use 'has-sitelib = false'",
             DeprecationWarning,
             stacklevel=2,
         )
@@ -119,7 +119,7 @@ def get_tag(pyproject) -> str:
     elif deprecate_build_system(pyproject, "any", False):
         warnings.warn(
             "The 'any = true' key is deprecated. "
-            "Please use 'has_sitelib = false' and 'has_binaries = false'",
+            "Please use 'has-sitelib = false' and 'has-binaries = false'",
             DeprecationWarning,
             stacklevel=2,
         )
@@ -127,14 +127,14 @@ def get_tag(pyproject) -> str:
     elif deprecate_build_system(pyproject, "pyver-any", False):
         warnings.warn(
             "The 'pyver-any = true' key is deprecated. "
-            "Please use 'has_binaries = false'",
+            "Please use 'has-binaries = false'",
             DeprecationWarning,
             stacklevel=2,
         )
         tag = f"py3{sys.version_info.minor}-none-any"
     else:
-        binaries = dotget(pyproject, "tool.cmeel.has_binaries", True)
-        sitelib = dotget(pyproject, "tool.cmeel.has_sitelib", True)
+        binaries = dotget(pyproject, "tool.cmeel.has-binaries", True)
+        sitelib = dotget(pyproject, "tool.cmeel.has-sitelib", True)
         if not binaries and not sitelib:
             tag = "py3-none-any"
         elif not binaries: